Split atlas overview and content ambient layouts
