+++ src/app/atlas/AtlasBlueprintHero.tsx
