+++ src/app/atlas/ambientConfig.ts
@@ -1,0 +1,7 @@
+export const ATLAS_BACKGROUND_CLASS = 'opacity-65';
+export const ATLAS_OVERLAY_CLASS =
+  'bg-[radial-gradient(circle_at_18%_22%,rgba(96,165,250,0.2),transparent_55%),radial-gradient(circle_at_82%_26%,rgba(244,114,182,0.18),transparent_55%),linear-gradient(140deg,rgba(2,6,23,0.95)_0%,rgba(10,21,38,0.92)_52%,rgba(15,23,42,0.94)_100%)]';
+
+export const ATLAS_BLUEPRINT_BACKGROUND_CLASS = 'opacity-65';
+export const ATLAS_BLUEPRINT_OVERLAY_CLASS =
+  'bg-[radial-gradient(circle_at_16%_20%,rgba(129,140,248,0.2),transparent_55%),radial-gradient(circle_at_86%_24%,rgba(192,132,252,0.22),transparent_60%),linear-gradient(150deg,rgba(2,6,23,0.95)_0%,rgba(11,26,48,0.92)_48%,rgba(15,23,42,0.94)_100%)]';
